requirements.txt
flask-oidc==1.4.0
okta==0.0.4
python-dotenv==0.10.1
pip>=18.1
setuptools>=39.0.1
